Fix OUTPUT_VALUE compiler warning
The issue involved the C language integer promotion rules.

+++ firmware/util.h
@@ -15,7 +15,7 @@
 
 #define FIND_MIDPOINT(min, max) ((min) + (((max) - (min)) / 2))
 
-#define OUTPUT_VALUE(v) do { PORTB = ~((uint8_t)(v << 6) | (uint8_t)(v >> 2)); } while(0)
+#define OUTPUT_VALUE(v) do { PORTB = (uint8_t)~((v << 6) | (v >> 2)); } while(0)
 
 #define SIZEOF_ARRAY(a) (sizeof(a) / sizeof(a[0]))
